+++ src/main/java/ru/platformer/game/model/Health.java
@@ -1,5 +1,5 @@
 package ru.platformer.game.model;
 
 public interface Health {
-    int getHealth();
+    float getHealth();
 }
